+++ src/model2/account_thresholds.ts
@@ -1,5 +1,3 @@
-import stellar from "stellar-base";
-
 export interface IAccountThresholds {
   masterWeight: number;
   low: number;
@@ -8,20 +6,6 @@ export interface IAccountThresholds {
 }
 
 export class AccountThresholds implements IAccountThresholds {
-  public static fromValue(value: string): AccountThresholds {
-    const t = Buffer.from(value, "base64");
-    const ti = stellar.xdr.ThresholdIndices;
-
-    const data: IAccountThresholds = {
-      masterWeight: t[ti.thresholdMasterWeight().value],
-      low: t[ti.thresholdLow().value],
-      medium: t[ti.thresholdMed().value],
-      high: t[ti.thresholdHigh().value]
-    };
-
-    return new AccountThresholds(data);
-  }
-
   public masterWeight: number;
   public low: number;
   public medium: number;
