+++ src/model2/factories/index.ts
@@ -1,1 +1,2 @@
 export * from "./account_flags_factory";
+export * from "./account_thresholds_factory";
